Rename to liwg-fs & minor server changes
